Simple shop UI (#4)
* Add global singleton for game state data

* Enabled the vim plugin in the addons/ dir

* Stock Market ui Widget

In the level ui scene, create and add a widget that ticks as new stock
values update.
Displays the graph, draws the lines, and shows the price value along
with a percent differential between the previous price value.

* Add test buttons to interact with the stock market

A "Buy Share" button to use funds to purchase a stock share.
A "Sell Share" button to sell a share of an owned stock and add funds
back to inventory.

* Create shop

Added a simple shop that displays information about a gun and an option
to purchase. The test button buys the weapon and should update the
bullet cost and equipped weapon label

* Shop UX/UI update plus refactor

The test buttons are functional. Buy a gun from the shop if funds are
available. The button is disabled when there aren't enough funds.

Refactor Items:
	- Collapse code blocks into simpler
	- Make guns Dictionary types rather than Array
	- Add detection for which gun fired to bill the player properly