Post-Level Summary Screen
Once the level is complete/failed/died, the level scene is destroyed and
a summary UI scene is created with a "You Win/Lose/Died" message at the
top.

The player's stats are calculated and billed accordingly.
The labels for each item are generated programmatically for most
flexibilty and appear in sequence on a timer.